Separate semver PR label check into its own workflow
Putting this into the Continuous Integration workflow caused the workflow to trigger two times when a PR is created because creating the PR and applying the first label seem to be both triggered. This was causing 9 build jobs to run twice unnecessarily.

This change separates out the semver PR label check into its own workflow. This semver PR label check workflow is still triggered twice when a PR is created with an initial label. At least the Continuous Integration workflow is not triggered twice.