chore: do not fail the CI build for low coverage on JRuby and TruffleRuby
There are many cases where JRuby and TruffleRuby
do not catch lines that were covered when they actually
were covered.

Instead, coverage will only be enforced for MRI Ruby on
both Linux and Windows.

Tests will still be run on JRuby and TruffleRuby and
coverage will be reported, but low coverage will NOT
fail the build.