feat!: rename ProcessExecuter.spawn_and_wait to spawn_with_timeout
This change is being made to feature the timeout functionality of this method.

BREAKING CHANGE: Users who use ProcessExecuter.spawn_and_wait will need to update
their calls to spawn_with_timeout. In addition, the following items will need to be
updated if used by the user of this gem:
* ProcessExecuter.spawn_and_wait_with_options
* ProcessExecuter::SpawnAndWaitOptions