feat!: remove Result#stdout and Result#stderr
Do not automatically capture stdout and stderr and return it in the Result object.

The way it was implemented, stdout and stderr were only captured if wrapped in a
MonitoredPipe. It also meant that every MonitoredPipe output destination was captured
even if it was not from stdout or stderr.

The new method will be added to the API `ProcessExecuter.run_with_capture` that
captures ONLY stdout and stderr and returns a `ResultWithCapture` which derives from
`Result` and has `#stdout` and `#stderr` methods.

BREAKING CHANGE: Users depending on `Result#stdout` or `Result#stderr` will either
have to capture this output manually themselves or change from `spawn_and_wait`/`run`
to `run_with_capture`.