feat: add `ProcessExecuter.run_with_capture`
Wraps `run` and automatically captures stdout and stderr

The captured output is accessed in the returned object's stdout and stderr
methods. Merged output (if the `merged_output: true` option is given) is
accessed in the stdout method.

stdout and stderr redirection options may be given by the caller. This will
override the capture if given. This means that if an stdout redirection is given,
the result.stdout will be empty and if a stderr redirection is given, the
result.stderr will be empty. A `ProcessExecuter::ArgumentError` will be raised if
both a stderr redirection and the `merge_output: true` option are given.

Accepts the same options as `run` and adds the following option:

* `:merge_output` to merge stdout and stderr into a single capture buffer (default is false)

A `ProcessExecuter::ArgumentError` will be raised if both an options object and
an options_hash are given.